requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
beautifulsoup4==4.13.1
certifi==2025.1.31
click==8.1.8
fastapi==0.115.8
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multidict==6.1.0
propcache==0.2.1
psutil==6.1.1
pydantic==2.10.6
pydantic_core==2.27.2
Pygments==2.19.1
python-dotenv==1.0.1
rich==13.9.4
sniffio==1.3.1
soupsieve==2.6
starlette==0.45.3
typing_extensions==4.12.2
uvicorn==0.34.0
yarl==1.18.3
